chore: packaging
- data, presentation 레이어 분리
