+++ app/src/main/AndroidManifest.xml
@@ -9,7 +9,7 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.GitSearcher">
-        <activity android:name=".SearcherActivity">
+        <activity android:name=".presentation.view.SearcherActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
